Added credential key to AuditActor
