+++ src/NuGetGallery/NuGetGallery.csproj
@@ -768,6 +768,7 @@
     <Compile Include="Controllers\PagesController.cs" />
     <Compile Include="Filters\ApiScopeRequiredAttribute.cs" />
     <Compile Include="GlobalSuppressions.cs" />
+    <Compile Include="Helpers\AuditingHelper.cs" />
     <Compile Include="Helpers\GravatarHelper.cs" />
     <Compile Include="Helpers\RegexEx.cs" />
     <Compile Include="Services\AsynchronousPackageValidationInitiator.cs" />
