+++ tests/NuGetGallery.Core.Facts/Auditing/AuditEntryTests.cs
@@ -27,6 +27,7 @@ namespace NuGetGallery.Auditing
                 machineIP: null,
                 userName: null,
                 authenticationType: null,
+                credentialKey: null,
                 timeStampUtc: DateTime.MinValue);
             var entry = new AuditEntry(record.Object, actor);
